==> analysis/R_analysis/scr/00_libraries.R <==
renv::install("clustree")
renv::install("pals")
renv::install("data.tree")
renv::install("HGNChelper")
renv::install("openxlsx")
renv::install("phipsonlab/speckle")
renv::install("statmod")
renv::install("ggalluvial")

renv::install(c("coda.base", "psych"))
renv::install("kharchenkolab/sccore")
renv::install("kharchenkolab/conos")
renv::install("kharchenkolab/cacoa")
