==> analysis/R_analysis/scr/00_preprocessing_SoupX.R <==
# AIM ---------------------------------------------------------------------
# the aim of this script is to:
# 1. transfer the minimal number of files from the source directory to the project folder.
# 2. run SoupX and generate the cleaned raw matrices.

# libraries ---------------------------------------------------------------
library(tidyverse)
library(SoupX)
library(DropletUtils)

# run the soupx processing ------------------------------------------------
# identify the samples
id_sample <- dir("../../data/cellranger7/output/")

# x <- id_sample[1]
# run the soupx processing
lapply(id_sample,function(x){
  # track the progress
  print(x)
  # define the location of the output of cellranger
  file <- paste0("../../data/cellranger7/output/",x,"/outs/")
  # Load data and estimate soup profile
  sc <- load10X(file)
  # Estimate rho
  sc <- autoEstCont(sc)
  # Clean the data
  out <- adjustCounts(sc,roundToInt = T)
  # save the data
  DropletUtils:::write10xCounts(paste0("../../data/SoupX_default_cellranger7/",x), out)
})

# test 01 failed sample ---------------------------------------------------
# I have followed the comment in the SoupX support page https://github.com/constantAmateur/SoupX/issues/60
# "If your data really does have 75% contamination, you probably shouldn't use it as that level of contamination likely indicates something went very wrong in the experiment. To check, I would look at the plot generated by autoEstCont. Does it have two peaks of roughly equal height, with one being around .75? If so, try setting your contamination to the location of the lower peak and proceeding with your analysis. The other thing I'd do is make extensive use of plotMarkerMap to see what the expression ratio to the soup looks like for a few genes that are commonly contamination. Without knowing your experiment it's hard to say what these are likely to be, but HB and IG genes usually work."

# based on the first plot generated from autoEstCont, I estimate a range of contamination between 0 and 0.2
# id_sample2 <- x <- "W8_CSF_tolebrutinib"
# 
# file <- paste0("../../data/cellranger7_out/",x)
# # Load data and estimate soup profile
# sc <- load10X(file)
# 
# # Estimate rho
# # fails
# sc2 <- autoEstCont(sc)
# # manual set rho
# sc3 <- setContaminationFraction(sc,contFrac = 0.1)
# # reduce the range for the estimate
# sc4 <- autoEstCont(sc,contaminationRange = c(0.01,0.5))
# 
# # plotMarkerMap(sc4,geneSet = "OLIG1") + plotMarkerMap(sc,geneSet = "OLIG1")
# out3 <- adjustCounts(sc3,roundToInt = T)
# out4 <- adjustCounts(sc4,roundToInt = T)
# 
# plotChangeMap(sc3, out3, "OLIG1") + plotChangeMap(sc4, out4, "OLIG1")
# plotChangeMap(sc3, out3, "SOX10") + plotChangeMap(sc4, out4, "SOX10")
# 
# remove(list = c("sc","sc3","sc4","out3","out4"))

# I eventually decided to use the autoEstCont(sc,contaminationRange = c(0.01,0.5)) only for this sample
lapply(id_sample2,function(x){
  # track the progress
  print(x)
  # define the location of the output of cellranger
  file <- paste0("../../data/cellranger7_out/",x)
  # Load data and estimate soup profile
  sc <- load10X(file)
  # Estimate rho
  sc <- autoEstCont(sc,contaminationRange = c(0.01,0.5))
  # Clean the data
  out <- adjustCounts(sc,roundToInt = T)
  # save the data
  DropletUtils:::write10xCounts(paste0("../../data/SoupX_default_cellranger7/",x), out)
})
